+++ test/strongbox/main_test.clj
@@ -26,8 +26,8 @@
 
   (testing "certain actions force the 'cli' ui"
     (is (= :cli (-> (main/parse ["--action" "scrape-catalogue"]) :options :ui)))
-    (is (= :cli (-> (main/parse ["--action" "scrape-curseforge-catalogue"]) :options :ui)))
-    (is (= :cli (-> (main/parse ["--action" "scrape-curseforge-catalogue"]) :options :ui))))
+    (is (= :cli (-> (main/parse ["--action" "scrape-github-catalogue"]) :options :ui)))
+    (is (= :cli (-> (main/parse ["--action" "scrape-github-catalogue"]) :options :ui))))
 
   (testing "certain actions force the 'cli' ui, even when 'gui' is explicitly passed"
     (is (= :cli (-> (main/parse ["--action" "scrape-catalogue" "--ui" "gui"]) :options :ui))))